Dan's idea 3 in practice
